changing packages to routes: only one package per directory
