+++ src/go.mod
@@ -2,6 +2,6 @@ module mstock-api/main
 
 go 1.14
 
-replace mstock-api/packages => ./packages
+replace mstock-api/routes => ./routes
 
-require mstock-api/packages v0.0.0
+require mstock-api/routes v0.0.0
